feat: refactor domain-related classes for improved structure and shared policies
