+++ src/main/java/br/com/jstack/org/catalog/graph/framework/adapter/input/rest/DomainRestController.java
@@ -1,11 +1,11 @@
 package br.com.jstack.org.catalog.graph.framework.adapter.input.rest;
 
 import br.com.jstack.org.catalog.graph.api.DomainsApi;
-import br.com.jstack.org.catalog.graph.application.usecase.CreateUseCase;
-import br.com.jstack.org.catalog.graph.application.usecase.DeleteByIdUseCase;
-import br.com.jstack.org.catalog.graph.application.usecase.RetrieveAllUseCase;
-import br.com.jstack.org.catalog.graph.application.usecase.RetrieveByIdUseCase;
-import br.com.jstack.org.catalog.graph.application.usecase.UpdateUseCase;
+import br.com.jstack.org.catalog.graph.application.port.input.shared.CreateUseCase;
+import br.com.jstack.org.catalog.graph.application.port.input.shared.DeleteByIdUseCase;
+import br.com.jstack.org.catalog.graph.application.port.input.shared.RetrieveAllUseCase;
+import br.com.jstack.org.catalog.graph.application.port.input.shared.RetrieveByIdUseCase;
+import br.com.jstack.org.catalog.graph.application.port.input.shared.UpdateUseCase;
 import br.com.jstack.org.catalog.graph.domain.aggregate.Domain;
 import br.com.jstack.org.catalog.graph.framework.adapter.mapper.DomainMapper;
 import br.com.jstack.org.catalog.graph.model.DomainRequest;
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
 @RestController
-public class DomainRestAdapter implements DomainsApi {
+public class DomainRestController implements DomainsApi {
 	
 	private final DomainMapper                        mapper;
 	private final CreateUseCase<Domain>               createUseCase;
